update script to load data from git repo
